Move PageTableToolbarFilter login into its own file
